Fully move python to cffi
